fix: improve project structure
