--- SpaceShip/docker-compose.debug.yml
@@ -1,18 +1,0 @@
-# Please refer https://aka.ms/HTTPSinContainer on how to setup an https developer certificate for your ASP.NET Core service.
-
-version: '3.4'
-
-services:
-  webapi:
-    image: webapi
-    build:
-      context: .
-      dockerfile: ./Dockerfile
-      args:
-        - configuration=Debug
-    ports:
-      - 5216:5216
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Development
-    volumes:
-      - ~/.vsdbg:/remote_debugger:rw
