--- SpaceShip/docker-compose.yml
@@ -1,24 +1,0 @@
-# Please refer https://aka.ms/HTTPSinContainer on how to setup an https developer certificate for your ASP.NET Core service.
-
-version: '3.4'
-
-services:
-  webapi:
-    image: webapi
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - 5216:5216
-    environment:
-      - RABBITMQ_HOST=rabbitmq-otus
-      - RABBITMQ_USER=space-ship
-      - RABBITMQ_PASSWORD=Ss1234
-      - RABBITMQ_TROUBLES_QUEUE=troubles
-      - RABBITMQ_STEP_QUEUE=steps
-    networks:
-      - otus-space
-
-networks:
-  otus-space:
-    name: otus-space
