+++ SpaceShip/src/SpaceShip.Service.EventsProvider/Interfaces/IEventConsumer.cs
