+++ SpaceShip/src/SpaceShip.Service.EventsProvider/SpaceShip.Service.EventsProvider.csproj
