+++ SpaceShip/src/SpaceShip.Service/Contracts/ResourceStateDTO.cs
@@ -1,0 +1,8 @@
+namespace SpaceShip.Service.Contracts;
+
+public enum ResourceStateDTO
+{
+    Sleep,
+    Normal,
+    Fail
+}
