--- SpaceShip/src/SpaceShip.Service/Models/Resorce.cs
@@ -1,1 +1,0 @@
-namespace SpaceShip.Service.Models;
