+++ SpaceShip/src/SpaceShip.WebAPI/Controllers/SpaceShipController.cs
@@ -1,13 +1,9 @@
-using System;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Components.Forms.Mapping;
+using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using AutoMapper;
 using SpaceShip.Service.Interfaces;
 using SpaceShip.WebAPI.Models;
+using System;
 
 namespace SpaceShip.WebAPI.Controllers;
