+++ SpaceShip/src/SpaceShip.WebAPI/Mappers/SpaceShipMappingProfile.cs
@@ -1,9 +1,6 @@
-using System.Security.Cryptography.X509Certificates;
 using AutoMapper;
-
 using SpaceShip.Service.Contracts;
 using SpaceShip.WebAPI.Models;
-
 
 namespace SpaceShip.WebAPI.Mappers;
 
@@ -12,8 +9,12 @@ namespace SpaceShip.WebAPI.Mappers;
 /// </summary>
 public class SpaceShipMappingProfile : Profile
 {
+    /// <summary>
+    /// Конструктор, при создании описываем настройки маппинга 
+    /// </summary>
     public SpaceShipMappingProfile()
     {
+        CreateMap<ResourceStateDTO, ResourceState>().ReverseMap();
         CreateMap<ResourceDTO, Resource>().ReverseMap();
         CreateMap<SpaceShipDTO,SpaceShipMetricResponse>().ReverseMap();
