+++ SpaceShip/src/SpaceShip.WebAPI/Models/SpaceShipCreateResponse.cs
