+++ SpaceShip/src/SpaceShip.WebAPI/Models/SpaceShipMetricResponse.cs
