+++ SpaceShip/src/SpaceShip.WebAPI/WebAPI.csproj
@@ -12,16 +12,16 @@
     <PackageReference Include="AutoMapper" Version="12.0.1" />
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
-    <PackageReference Include="RabbitMQ.Client" Version="6.8.1" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.6.2" />
+    <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="6.6.2" />
     <PackageReference Include="System.Threading" Version="4.3.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SpaseShip.Service.EventsProvider\SpaseShip.Service.EventsProvider.csproj" />
     <ProjectReference Include="..\..\mock\MockSpaceShip.Service\MockSpaceShip.Service.csproj" />
     <ProjectReference Include="..\SpaceShip.Service\SpaceShip.Service.csproj" />
+    <ProjectReference Include="..\SpaceShip.Service.EventsProvider\SpaceShip.Service.EventsProvider.csproj" />
   </ItemGroup>
 
 </Project>
