--- SpaceShip/src/WebAPI/Settings/ApplicationSettings.cs
@@ -1,12 +1,0 @@
-#pragma warning disable CS1591
-
-namespace SpaceShip.WebAPI.Settings;
-
-/// <summary>
-/// Сериализация настроек приложения
-/// </summary>
-public class ApplicationSettings: SettingsBase
-{
-  
-
-}
