--- SpaceShip/src/WebAPI/Settings/PostgresConnectionSettings.cs
@@ -1,9 +1,0 @@
-#pragma warning disable CS1591
-
-namespace SpaceShip.WebAPI.Settings;
-
-public class PostgresConnectionSettings : SettingsBase
-{
-    public string? PostgresConnectionString {get; set;}
-
-}
